backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
torch==2.5.0
torchvision==0.20.0
transformers==4.45.0
scikit-learn==1.5.2
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
aiofiles==24.1.0
setuptools==75.1.0
reportlab==4.4.4
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
geoalchemy2==0.18.0
alembic==1.16.5
ultralytics==8.3.66
segmentation-models-pytorch==0.3.4
